feat(filters): Load books from search with price and type filters
