+++ src/view/Search/index.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect } from "react";
 import Header from "../../components/header";
 import Filters from "../../components/filters";
 import {
@@ -17,33 +17,49 @@ import {
   privacyAndPolicy,
   termsAndUsage,
 } from "../../constants";
-import { books } from "../../components/shelves/constants";
+import { useCachedfetch } from "../../hooks/useCachedFetch";
+
 
 const Search: React.FC = () => {
+  const {data, load} = useCachedfetch<any>();
+
+  useEffect(() => {
+    const query = new URLSearchParams(window.location.search);
+    if(!query.size) return;
+    
+    const apiUrl = new URL(`https://www.googleapis.com/books/v1/volumes?&startIndex=0&maxResults=40`)
+    apiUrl.search = `${apiUrl.search}&${query.toString()}`;
+    
+    load(apiUrl.href);
+  }, []);
+
+
   return (
     <>
       <Header />
       <Container>
         <Content>
-        <Filters mainTitle="Filter" hasSelectedFilters={false} resetFilters={() => {}} />
+        <Filters mainTitle="Filter"/>
         <ContentResults>
-          {books.map((shelf) => (
-            <>
-              {shelf.booksShelf.map((book) => (
-                <ContentResultsWrapper>
-                  <ContentResultsCover>
-                    <img src={book.urlImage} alt={book.slug} />
-                  </ContentResultsCover>
-                  <ContentResultsTitle>
-                    <label>{book.title} </label>
-                  </ContentResultsTitle>
-                  <ContentResultsCategory>
-                    <span>{book.autor}</span>
-                  </ContentResultsCategory>
-                </ContentResultsWrapper>
-              ))}
-            </>
-          ))}
+          
+          <>
+            {data?.items.map((book: any) => (
+              <ContentResultsWrapper key={book.id}>
+                <ContentResultsCover>
+                  <img
+                    src={book.volumeInfo.imageLinks?.thumbnail || `https://placehold.co/85x124?text=${book.volumeInfo.title}`}
+                    alt={book.volumeInfo.title} />
+                </ContentResultsCover>
+                <ContentResultsTitle>
+                  <label>{book.volumeInfo.title} </label>
+                </ContentResultsTitle>
+                <ContentResultsCategory>
+                  <span>{book.volumeInfo.authors}</span>
+                </ContentResultsCategory>
+              </ContentResultsWrapper>
+            ))}
+          </>
+          
         </ContentResults>
         </Content>
       </Container>
